move: getJSONSchema
