+++ vite.config.ts
@@ -12,9 +12,9 @@ import { type Connect, defineConfig, type PluginOption, type ViteDevServer } fro
 import checker from 'vite-plugin-checker';
 import tsconfigPaths from 'vite-tsconfig-paths';
 import moduleJSON from './module.json' with { type: 'json' };
+import { getJSONSchema } from './scripts/buildJSONSchema';
 import { type FileValidationFailure, Log, pluralise } from './scripts/helpers';
 import { testAndMergeAnimations } from './scripts/testAndMergeAnimations';
-import { getJSONSchema } from './src/storage/animationsSchema';
 
 const packagePath = `modules/${moduleJSON.id}`;
